test: change order of e2e tests again to debug CI
